@@ -37,26 +37,24 @@
     "@babel/core": "^7.14.6",
     "@babel/preset-env": "^7.14.7",
     "@babel/preset-typescript": "^7.14.5",
-    "@commitlint/cli": "^13.1.0",
-    "@commitlint/config-conventional": "^13.1.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/vue-fontawesome": "^2.0.2",
+    "@commitlint/cli": "^13.2.0",
+    "@commitlint/config-conventional": "^13.2.0",
     "@rollup/plugin-alias": "^3.1.2",
     "@rollup/plugin-babel": "^5.3.0",
     "@rollup/plugin-commonjs": "^20.0.0",
     "@rollup/plugin-node-resolve": "^13.0.5",
     "@rollup/plugin-replace": "^3.0.0",
-    "@storybook/addon-actions": "^6.3.8",
-    "@storybook/addon-essentials": "^6.3.8",
-    "@storybook/addon-links": "^6.3.8",
+    "@storybook/addon-actions": "^6.3.9",
+    "@storybook/addon-essentials": "^6.3.9",
+    "@storybook/addon-links": "^6.3.9",
     "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/vue": "^6.3.8",
-    "@typescript-eslint/eslint-plugin": "^4.31.2",
-    "@typescript-eslint/parser": "^4.31.2",
+    "@storybook/vue3": "^6.3.9",
+    "@typescript-eslint/eslint-plugin": "^4.32.0",
+    "@typescript-eslint/parser": "^4.32.0",
     "@vue/cli-plugin-babel": "^4.5.13",
     "@vue/cli-plugin-typescript": "^4.5.13",
     "@vue/cli-service": "^4.5.13",
+    "@vue/compiler-sfc": "^3.0.11",
     "@zerollup/ts-transform-paths": "^1.7.18",
     "babel-loader": "^8.2.2",
     "can-npm-publish": "^1.3.6",
@@ -72,39 +70,41 @@
     "husky": "^7.0.2",
     "lint-staged": "^11.1.2",
     "minimist": "^1.2.5",
+    "postcss": "^8.2.10",
     "prettier": "^2.4.1",
     "rimraf": "^3.0.2",
     "rollup": "^2.57.0",
+    "rollup-plugin-postcss": "^4.0.0",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-typescript2": "^0.30.0",
-    "rollup-plugin-vue": "^5.1.9",
+    "rollup-plugin-vue": "^6.0.0",
     "sass": "^1.42.1",
     "sass-loader": "^10.2.0",
     "style-loader": "^2.0.0",
     "stylelint": "^13.13.1",
+    "stylelint-config-prettier": "^8.0.2",
     "stylelint-config-recess-order": "^2.5.0",
     "stylelint-config-standard": "^22.0.0",
     "ttypescript": "^1.5.12",
     "typescript": "^4.4.3",
-    "vue": "^2.6.14",
-    "vue-loader": "^15.9.8",
-    "vue-template-compiler": "^2.6.14"
+    "vue": "^3.0.5",
+    "vue-loader": "^16.8.1"
   },
   "peerDependencies": {
-    "vue": "^2.6.14"
+    "vue": "^3.0.5"
   },
   "engines": {
-    "node": ">=12"
+    "node": ">=14"
   },
   "scripts": {
+    "prepare": "husky install",
     "serve": "cross-env PORT=3000 vue-cli-service serve dev/serve.ts",
-    "lint": "npm run lint:js & npm run lint:style",
-    "lint:js": "eslint . --ext .js,.ts,.vue",
+    "lint": "npm run lint:script & npm run lint:style",
+    "lint:script": "eslint . --ext .js,.jsx,.ts,.tsx,.vue",
     "lint:style": "stylelint \"**/*.{css,scss,vue}\"",
-    "format": "prettier --write \"**/*.{js,ts,vue,json}\"",
+    "format": "prettier \"**/*.{js,jsx,ts,tsx,vue,json}\" --write",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
-    "prepare": "husky install",
     "prebuild": "rimraf ./dist",
     "build": "cross-env NODE_ENV=production rollup --config build/rollup.config.js",
     "build:ssr": "cross-env NODE_ENV=production rollup --config build/rollup.config.js --format cjs",
