@@ -4,13 +4,13 @@ module.exports = {
     es2021: true,
     node: true,
   },
-  extends: ['plugin:vue/recommended', 'standard', 'prettier'],
+  extends: ['plugin:vue/vue3-recommended', 'standard', 'prettier'],
   parserOptions: {
     ecmaVersion: 12,
     parser: '@typescript-eslint/parser',
     sourceType: 'module',
   },
-  plugins: ['vue', '@typescript-eslint'],
+  plugins: ['@typescript-eslint'],
   rules: {
     'vue/component-tags-order': [
       'error',
